requirements.txt
Flask==2.2.3
pandas==1.5.3
scikit-learn==1.2.1
pymysql==1.0.2
python-dotenv==1.0.0
